Update technotes to use documenteer 0.2 series
documenteer 0.2.x changes how technotes are configured. It also reduces
the number of Python packages that must be managed to a technote to just
documenteer[technote]; documenteer provides sphinx, sphinx extensions,
and the sphinx theme.

With documenteer 0.2 we leave intersphinx blank for performance. This
commit adds an example in conf.py to help people configure it if needed.